Player, Camera, and Cloud Movement
Created scene with a "Ground" starting platform.

Created player with movement (needs more testing + finetuning but fully functional). Player is a rocket that starts with 100 energy. They can use A and D or the left/right arrow keys to tilt the rocket left/right and space bar or LMB to boost, consuming energy to fly the player towards their relative up direction. The player is locked in screen space by colliders on either side of the cinemachine camera.

The camera is a cinemachine camera with a position composer set to follow the player. The camera is locked on the X axis with a very simple script, CameraLockX.cs, which simply sets the attached GameObject's X position to 0 in LateUpdate(), after any camera movement that would be handled in Update(). It has two very tall colliders, one on either side, which move with the camera to keep the player inside the camera's width.

Clouds' only current functionality is to move left/right and collide with (but not move) the player. The cloud prefab has a box collider with a platform effector and is spawned by the CloudManager.cs script, which currently just spawns a given number of clouds evenly spaced starting at Y = 6, randomly choosing to spawn each cloud on the far left/right. Cloud.cs picks a random movement speed within a set range (currently 1-2), then moves the cloud left/right, changing direction when reaching the edge of the screen.