Created Temp Main Menu and Game Over Bug Fix
Created temporary basic main menu. Used the icon for the jam as the background and made a basic play button and info button with an info pop-up explaining the goal of the game and listing images and names for the different kinds of clouds and batteries. Images for clouds and batteries are just screenshots taken from the unity editor and buttons just use basic default unity shapes and colors.

Bug Fix: Moved death logic when colliding with platform under the check for if the player is above the platform so that the player doesn't lose when using all their boost to fly high really fast.

Built V0.9 off of this version and uploaded to itch.io